Simplifying statespace: Moved NavigateActionType to nav package and added LocationParameterizable interface
