+++ src/taxi/functions/amdp/NavigateAbstractPF.java
@@ -3,8 +3,8 @@ package taxi.functions.amdp;
 import burlap.mdp.core.oo.propositional.PropositionalFunction;
 import burlap.mdp.core.oo.state.OOState;
 import taxi.Taxi;
-import taxi.hierarchies.tasks.put.NavigateActionType;
-import taxi.hierarchies.tasks.put.NavigateActionType.NavigateAction;
+import taxi.hierarchies.tasks.nav.NavigateActionType;
+import taxi.hierarchies.tasks.nav.NavigateActionType.NavigateAction;
 import taxi.hierarchies.tasks.nav.state.TaxiNavState;
 
 public class NavigateAbstractPF extends PropositionalFunction {
