+++ src/taxi/hierarchies/interfaces/LocationParameterizable.java
@@ -1,0 +1,5 @@
+package taxi.hierarchies.interfaces;
+
+public interface LocationParameterizable {
+    String[] getLocations();
+}
