+++ src/taxi/hierarchies/tasks/get/state/TaxiGetState.java
@@ -7,9 +7,10 @@ import burlap.mdp.core.oo.state.OOStateUtilities;
 import burlap.mdp.core.oo.state.OOVariableKey;
 import burlap.mdp.core.oo.state.ObjectInstance;
 import burlap.mdp.core.state.MutableState;
+import taxi.hierarchies.interfaces.LocationParameterizable;
 import taxi.hierarchies.tasks.get.TaxiGetDomain;
 
-public class TaxiGetState implements MutableOOState {
+public class TaxiGetState implements MutableOOState, LocationParameterizable {
 
 	//this state has passengers and depots
 	private TaxiGetAgent taxi;
